+++ crates/azure/src/auth.rs
@@ -1,4 +1,4 @@
-use cloud_terrastodon_azure_types::prelude::User;
+use cloud_terrastodon_azure_types::prelude::EntraUser;
 use cloud_terrastodon_command::CacheKey;
 use cloud_terrastodon_command::CommandBuilder;
 use cloud_terrastodon_command::CommandKind;
@@ -6,7 +6,7 @@ use eyre::Result;
 use std::path::PathBuf;
 use tracing::warn;
 
-pub async fn fetch_current_user() -> Result<User> {
+pub async fn fetch_current_user() -> Result<EntraUser> {
     CommandBuilder::new(CommandKind::AzureCLI)
         .args(["ad", "signed-in-user", "show"])
         .cache(CacheKey::new(PathBuf::from_iter([